Development (#18)
* 2.1.3 :: Add documentarion / change this icons.

* Remove news extension dependency

* Removes the news extension dependecy (Removes the newsletter task. You
can use the scheduler task cy_newsletter.)

* Update bootstrap dependency

* Update bootstrap_package dependency: Allows the version 13.

* Update readme.md

* Update readme.md

* update readme.md

* Repair the plugin configuration / registry.

* UPD : Update to TYPO3 12.4.0

* Update to bootstrap_package 14.0.0

* 3.0.3 :: FIX : The confirm new email address change mail is send to the current and to the future email address.

* Update to TYPO3 13.4.0

* UPD : Update (cleaning) TCA configuration

* 4.1.3 :: CLN : Cleaning fe_users integration

* 4.1.4 :: FIX : The passwort change form works as expected.

* 4.2.1 :: FIX : Email change tutorial fixed.
4.2.0 :: ADD : When editing the profile, you can restrict the user's right to change their image.

* 4.2.1 :: FIX : Email change tutorial fixed.

---------

+++ Classes/Controller/ProfileController.php
@@ -34,6 +34,7 @@ class ProfileController extends ActionController
 
 
     private const CURRENT_USER = 'currentUser';
+    private const USER_CAN_CHANGE_IMAGE = 'userCanChangeImage';
 
     private const EMAIL = 'email';
 
@@ -105,9 +106,9 @@ class ProfileController extends ActionController
         }
 
         if (!$validationResults->hasErrors()) {
-
-            $this->updatePortrait($currentUser);
-
+            if ($this->settings['userCanChangeImage'] ?? true) {
+                $this->updatePortrait($currentUser);
+            }
             $this->frontendUserRepository->update($currentUser);
             $validationResults->addInfo('savingSuccessful');
         }
@@ -147,19 +148,21 @@ class ProfileController extends ActionController
     {
         /** @var ValidationResults $validationResults **/
         $validationResults = $this->getValidationResults();
+        if ($this->settings['userCanChangeImage'] ?? true) {
 
-        if (!$this->frontendUserService->isLogged()) {
-            $validationResults->addError('notLogged');
-        }
-        $frontendUser = $this->frontendUserService->getCurrentUser();
-        $portrait = $frontendUser->getPortrait();
-        if (!$validationResults->hasErrors() && $portrait != null) {
+            if (!$this->frontendUserService->isLogged()) {
+                $validationResults->addError('notLogged');
+            }
+            $frontendUser = $this->frontendUserService->getCurrentUser();
+            $portrait = $frontendUser->getPortrait();
+            if (!$validationResults->hasErrors() && $portrait != null) {
 
-            $frontendUser->setPortrait(null);
-            $fileResource = $this->resourceFactory->getFileReferenceObject($portrait->getUid());
-            $fileResource->getOriginalFile()->delete();
+                $frontendUser->setPortrait(null);
+                $fileResource = $this->resourceFactory->getFileReferenceObject($portrait->getUid());
+                $fileResource->getOriginalFile()->delete();
 
-            $this->frontendUserRepository->update($frontendUser);
+                $this->frontendUserRepository->update($frontendUser);
+            }
         }
         return GeneralUtility::makeInstance(ForwardResponse::class, 'editProfile')->withArguments([
             ProfileController::VALIDATIOPN_RESULTS => $validationResults
@@ -179,7 +182,7 @@ class ProfileController extends ActionController
             $frontendUser = $this->frontendUserService->getCurrentUser();
             $this->view->assign(ProfileController::CURRENT_USER, $frontendUser);
             $this->view->assign(ProfileController::VISIBLE_FE_GROUPS, GeneralUtility::intExplode(',', $this->settings['visibleFeGroups']));
-
+            $this->view->assign(ProfileController::USER_CAN_CHANGE_IMAGE, $this->settings['userCanChangeImage'] ?? true);
 
             // migration from usertools 3.x (for TYPO3 12) to 4.x (for TYPO3 13)
             // Moves the first image to the portait, if the filename is equals with the username.
